+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [[v0.3.1]](https://github.com/mlange-42/arche-model/compare/v0.3.0...v0.3.1)
+
+### Other
+
+* Increased time precision on Windows for more consistent TPS and FPS (#39)
+
 ## [[v0.3.0]](https://github.com/mlange-42/arche-model/compare/v0.2.0...v0.3.0)
 
 ### Breaking changes